Checkout tags to local branches when making multi-version docs. (#9)
The previous approach of checking out detached HEAD led to unexpected build 
failures when building the API documentation for multiple versions of the code. 
We now create a local branch and delete it when creating the docs for each version.
We also include a check for `git branch --show-current` returning an empty string.