Backwards compatibility for scipy 1.5.
Scipy 1.6.0 introduced the multivariate_t random variable, which we use
in our implementation of conjugate priors. However, scipy 1.6.0+
requires Python 3.7+. To maintain backwards compatibility with Python
3.6 (and therefore scipy 1.5), we implement the log density of the
multivariate t distribution & use it as a fallback where necessary.

We also implement an optimized computaiton of pseudo-inverse, and
explicitly allow for singular V_0 matrices in the computation of the
Bayesian Multivariate Linear Regression posterior.

Finally, we update the testing workflows to avoid segfaults due to BLAS
bugs in older package versions.