Initial pass at fixing autoML implementation.
The existing implementation of autoML layers is not compatible with the
ensemble framework. This commit takes a first pass at unifying the two.
In particular, we define a new layer config object which holds a model
directly. It is likely that we will re-implement ensembles in the same
way, i.e. the ensemble config will hold the models directly, instead of
the ensemble model holding the models & the config holding model
configs.