Exception handling for ensembles and Spark API (#118)
* Better version of predict_on_train for Spark.

Previously, we called model.forecast(ts, prev, return_prev=True). Now,
we just use the return value of model.train() for the predictions on the
train data, if desired.

* Add exception handling for ensembles.

Previously ensembles would error out if any of the individual models
threw an exception during training or evaluation. Now, these models will
just not be used.

* Add exception handling in pyspark API.

If an exception occurs when training a model or using it for inference,
we now provide some default predictions as a fallback. This improves
robustness.

* Add reset() method for ensemble combiners.

* Retain some state when calling combiner.reset().

Re-training the combiner will automatically refresh the number of models
(if needed) and the values of an evaluation metric (if validation data
is given). However, some unexpected behavior can arise if these are
always reset to None.

* Add test fixture for spark session.