Improvements to AutoML forecasting detectors.
Many AutoML forecasters return train_pred, train_err when trained.
However, when they are used for anomaly detection, training is expected
to return the sequence of training anomaly scores. This commit creates a
new train_post_process() helper method which allows the gap to be
bridged with a single line of code.