More robust detection of time series granularity.
Previously, we would detect the granularity of a time series as the GCD
of all timedeltas found in the time series (assuming pandas couldn't
infer the granularity on its own). However, this behavior fails for
time series with missing data that are sampled at granularities that
aren't an exact number of seconds, e.g. monthly time series would be
resampled to a daily granularity because months are of inconsistent
length.

This commit uses the most commonly observed timedelta, and it also
checks whether a k-month granularity is a better fit for the time series
than a n-day granularity.